requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pydantic>=2.0.0
python-dotenv>=1.0.0